[REFACTOR/#317] 알림 권한 로직 개선
- 알림 권한 요청 로직 삭제 (TimeReminderScreen)
- Home 화면 진입 시 알림 권한 확인하도록 수정 (HomeScreen)
